refactor: extract feature seams from oversized files to satisfy the changed-file LOC ratchet
