+++ packages/gateway-protocol/src/index.ts
@@ -4,6 +4,7 @@ export { validateApprovalResolveResult } from "./approval-result-validators.js";
 import type { ValidationError } from "./validation-errors.js";
 export { formatValidationErrors, type ValidationError } from "./validation-errors.js";
 import { lazyCompile } from "./protocol-validator.js";
+import { checkWorkerProtocolJson } from "./worker-protocol-json.js";
 export type { ProtocolValidator } from "./protocol-validator.js";
 export * from "./schema/worker-inference.js";
 export * from "./schema/skill-history.js";
@@ -96,7 +97,6 @@ import {
   ChatMessageGetResultSchema,
   ChatMessageGetParamsSchema,
   ChatInjectParamsSchema,
-  type ChatInjectBashExecutionParams,
   ChatInjectBashExecutionParamsSchema,
   ChatSendParamsSchema,
   ChatToolTitlesParamsSchema,
@@ -486,51 +486,6 @@ export const validateConnectParams = lazyCompile(ConnectParamsSchema);
 export const validateWorkerAdmissionHandshake = lazyCompile(WorkerAdmissionHandshakeSchema);
 export const validateWorkerConnectRequestFrame = lazyCompile(WorkerConnectRequestFrameSchema);
 export const validateWorkerHeartbeatParams = lazyCompile(WorkerHeartbeatParamsSchema);
-
-function checkWorkerProtocolJson(data: unknown): ValidationError | undefined {
-  const stack: Array<{ depth: number; value: unknown }> = [{ depth: 0, value: data }];
-  const seen = new WeakSet<object>();
-  while (stack.length > 0) {
-    const current = stack.pop();
-    if (!current) {
-      break;
-    }
-    if (current.depth > WORKER_TRANSCRIPT_MAX_JSON_DEPTH) {
-      return {
-        keyword: "maxDepth",
-        params: { limit: WORKER_TRANSCRIPT_MAX_JSON_DEPTH },
-        message: `must not exceed JSON nesting depth ${WORKER_TRANSCRIPT_MAX_JSON_DEPTH}`,
-      };
-    }
-    if (
-      current.value === null ||
-      typeof current.value === "string" ||
-      typeof current.value === "boolean"
-    ) {
-      continue;
-    }
-    if (typeof current.value === "number") {
-      if (!Number.isFinite(current.value)) {
-        return { keyword: "finite", message: "must contain only finite JSON numbers" };
-      }
-      continue;
-    }
-    if (typeof current.value !== "object") {
-      return { keyword: "jsonValue", message: "must contain only JSON values" };
-    }
-    if (seen.has(current.value)) {
-      return { keyword: "acyclic", message: "must be an acyclic JSON value" };
-    }
-    seen.add(current.value);
-    const values = Array.isArray(current.value)
-      ? current.value
-      : Object.values(current.value as Record<string, unknown>);
-    for (const value of values) {
-      stack.push({ depth: current.depth + 1, value });
-    }
-  }
-  return undefined;
-}
 
 export const validateWorkerTranscriptCommitParams = lazyCompile(
   WorkerTranscriptCommitParamsSchema,
